backend/requirements.txt
langchain
langchain-community
langchain-google-genai
pandas
matplotlib
python-dotenv
# Backend API
fastapi
uvicorn
pydantic
